+++ apps/api/src/subscription/index.ts
@@ -1,3 +1,2 @@
-export { SubscriptionFacade } from "./application/facades/subscription.facade";
 export type { SubscriptionEventPayload } from "./application/types/subscription-event.payload";
 export { SubscriptionModule } from "./subscription.module";
